% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dsa.R
\name{plot_dsa}
\alias{plot_dsa}
\title{One-Way Deterministic Sensitivity Plot}
\usage{
plot_dsa(df, metric = "ICER")
}
\arguments{
\item{df}{A data frame from \code{\link[=dsa_icer]{dsa_icer()}}.}

\item{metric}{Character string. Either \code{"ICER"} or \code{"INMB"}.}
}
\value{
A \code{ggplot} object.
}
\description{
Creates a one-way sensitivity plot from a data frame generated by
\code{\link[=dsa_icer]{dsa_icer()}}.
}
\examples{
df <- simulate_ce_trial(n = 100, seed = 123)
dsa <- dsa_icer(cost + effect ~ group, data = df, param = "effect",
                range = seq(0.74, 0.82, 0.02), ref = "control")
plot_dsa(dsa, metric = "ICER")
}
