% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icer.R
\name{cea}
\alias{cea}
\title{Observed cost-effectiveness outcomes}
\usage{
cea(cost, effect, group, ref, data, na.omit = TRUE)
}
\arguments{
\item{na.omit}{}
}
\value{

}
\description{
Observed cost-effectiveness outcomes
}
\examples{
trt_grp <- data.frame(group = rep("treatment", 250), cost = rgamma(250, 1000), effect = runif(250, 1, 20))
control_grp <- data.frame(cost = rgamma(250, 10), effect = runif(250, 50, 100), group = rep("control", 250))
data <- rbind(trt_grp, control_grp)
ref <- 'control'

}
