% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tornado.R
\name{plot_tornado}
\alias{plot_tornado}
\title{Tornado Plot for Deterministic Sensitivity Analysis}
\usage{
plot_tornado(df, metric = "ICER")
}
\arguments{
\item{df}{A data frame from \code{dsa_icer()}.}

\item{metric}{Character string. Either "ICER" or "NMB". Must match the column name in \code{df}.}
}
\value{
A ggplot2 tornado-style line plot.
}
\description{
Creates a tornado plot from a data frame generated by \code{\link[=dsa_icer]{dsa_icer()}}.
}
\examples{
plot_tornado(dsa_result, metric = "ICER")
}
